Address review feedback on ownership percentage
- Save the owner's percentage in the same transaction as the members',
  so a rejected member value no longer leaves the owner's change persisted
- Assign a submitted member percentage whenever the field is present, so a
  blank value is rejected by the model instead of being silently ignored
  or falling back to the database default
- Keep sub-second precision in the share cache version of the net worth
  series builders so two edits within one second don't show a stale chart

The test environment uses a null cache store, so the cache test swaps in a
memory store to exercise caching for real.

+++ app/models/balance_sheet/net_worth_breakdown_series_builder.rb
@@ -132,7 +132,7 @@ class BalanceSheet::NetWorthBreakdownSeriesBuilder
     end
 
     def cache_key(period)
-      shares_version = user ? AccountShare.where(user: user).maximum(:updated_at)&.to_i : nil
+      shares_version = user ? AccountShare.where(user: user).maximum(:updated_at)&.to_f : nil
       key = [
         "balance_sheet_net_worth_breakdown_series",
         CACHE_VERSION,
