+++ app/models/balance_sheet/net_worth_series_builder.rb
@@ -44,7 +44,7 @@ class BalanceSheet::NetWorthSeriesBuilder
     end
 
     def cache_key(period)
-      shares_version = user ? AccountShare.where(user: user).maximum(:updated_at)&.to_i : nil
+      shares_version = user ? AccountShare.where(user: user).maximum(:updated_at)&.to_f : nil
       key = [
         "balance_sheet_net_worth_series_historical",
         user&.id,
